feat(mobile): add family profile lens tab
