+++ apps/mobile/src/hooks/use-family-presence.test.ts
@@ -1,0 +1,48 @@
+import { renderHook } from '@testing-library/react-native';
+
+import { useFamilyPresence } from './use-family-presence';
+
+const mockUseDashboard = jest.fn();
+
+jest.mock('./use-dashboard', () => ({
+  useDashboard: () => mockUseDashboard(),
+}));
+
+describe('useFamilyPresence', () => {
+  beforeEach(() => {
+    mockUseDashboard.mockReset();
+  });
+
+  it('returns hasFamily=false while loading', () => {
+    mockUseDashboard.mockReturnValue({ data: undefined, isLoading: true });
+    const { result } = renderHook(() => useFamilyPresence());
+    expect(result.current).toEqual({ hasFamily: false, isLoading: true });
+  });
+
+  it('returns hasFamily=false when children list is empty', () => {
+    mockUseDashboard.mockReturnValue({
+      data: { children: [], demoMode: false },
+      isLoading: false,
+    });
+    const { result } = renderHook(() => useFamilyPresence());
+    expect(result.current).toEqual({ hasFamily: false, isLoading: false });
+  });
+
+  it('returns hasFamily=true when at least one child is linked', () => {
+    mockUseDashboard.mockReturnValue({
+      data: { children: [{ profileId: 'c1' }], demoMode: false },
+      isLoading: false,
+    });
+    const { result } = renderHook(() => useFamilyPresence());
+    expect(result.current).toEqual({ hasFamily: true, isLoading: false });
+  });
+
+  it('treats demoMode children as real', () => {
+    mockUseDashboard.mockReturnValue({
+      data: { children: [{ profileId: 'demo' }], demoMode: true },
+      isLoading: false,
+    });
+    const { result } = renderHook(() => useFamilyPresence());
+    expect(result.current.hasFamily).toBe(true);
+  });
+});
