+++ apps/mobile/src/lib/navigation.test.ts
@@ -11,7 +11,7 @@ import type { Router } from 'expo-router';
 
 describe('navigation constants', () => {
   it('exports FAMILY_HOME_PATH for family-facing navigation', () => {
-    expect(FAMILY_HOME_PATH).toBe('/(app)/dashboard');
+    expect(FAMILY_HOME_PATH).toBe('/(app)/family');
   });
 });
